+++ src/Proyecto/movimiento.java
@@ -6,20 +6,20 @@ public class movimiento implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private String nombre;
-    private int potencia;
+    private double potencia;
     private String tipo;
     private int nivelRequerido;
     private Efectos efecto;
 
 
-    public movimiento(String nombre, int potencia, String tipo, int nivelRequerido, Efectos efecto) {
+    public movimiento(String nombre, double potencia, String tipo, int nivelRequerido, Efectos efecto) {
         this.nombre = nombre;
         this.potencia = potencia;
         this.tipo = tipo;
         this.nivelRequerido = nivelRequerido;
         this.efecto = efecto;
     }
-    public movimiento(String nombre, int potencia, String tipo, int nivelRequerido) {
+    public movimiento(String nombre, double potencia, String tipo, int nivelRequerido) {
         this.nombre = nombre;
         this.potencia = potencia;
         this.tipo = tipo;
@@ -30,7 +30,7 @@ public class movimiento implements Serializable {
             return nombre;
         }
 
-    public int getPotencia() {
+    public double getPotencia() {
             return potencia;
         }
